create a project with html5/css3 and js
HTML Structure:

It forms the layout of the calculator through the emulating HTML codes.
A div has the class “calculator” which makes it act as the main container.
Inside it, there will be a div with a class name “display”, where the input and result can be displayed.
The set of calculator buttons will be in the div with class “buttons”.
CSS Styling:

The CSS formats the calculator with nice lines that make the calculator look more sleek of the modern world.
It is used to define the background color, padding, border radius and the box shadow of the main calculator div.
It can adjust the aesthetics of display area, buttons as well as normal buttons, operator buttons, equal button, and the clear button.
JavaScript Functionality:

JavaScript is used in this calculator and it is responsible for the calculations being made.
It employs such variable as currentInput, which holds the current number or expression being worked on.
appendNumber(number): Adds a number to the currentInput string and changes the display of it.
appendOperator(operator): Adds an operator to the currentInput string and the displays the new string.
calculate(): To check the result the script uses the eval() function to evaluate the currentInput expression and prints it on the page.
clearDisplay(): Clears the currentInput value to “0” and clears the display of the calulator.
To use the code:

Now you need to make an HTML file (for example, calculator. html) and open it in the notepad to insert the above code into the file.
Load the HTML file in any web browser. You should see the calculator interface If not, then please, start the server and open it at http://127.0.0.1:3000.
It contains buttons which can be used to input numbers, mathematical operators and also the result of the computation.
It is simple Push button calculator’s code which has given basic functionality of such a calculator. You can expand it further to include more advanced features like:You can expand it further to include more advanced features like:

Memory functions: Save and retrieve earlier results and bring them into subsequent processes.
Trigonometric functions: Organise mathematical functions such as sine, cosine and tangent functions.
Scientific notation: Harps on the care of very big or very small numbers.
Error handling: Disposing or anticipating wrong input or mathematical errors.
Styling: Choose theme and buttons from various options that are available to make a website look different from the other ones.